Authenticaion fix

+++ app/Http/Controllers/PostController.php
@@ -19,6 +19,9 @@ class PostController extends Controller
 
     public function getAdminIndex()
     {
+        if(!Auth::check()){
+            return redirect()->back();
+        }
         $posts = Post::orderBy('title', 'asc')->get();
         return view('admin.index', ['posts' => $posts]);
     }
@@ -39,12 +42,18 @@ class PostController extends Controller
 
     public function getAdminCreate()
     {
+        if(!Auth::check()){
+            return redirect()->back();
+        }
         $tags = Tag::all();
         return view('admin.create', ['tags' => $tags]);
     }
 
     public function getAdminEdit($id)
     {
+        if(!Auth::check()){
+            return redirect()->back();
+        }
         $post = Post::find($id);
         $tags = Tag::all();
         return view('admin.edit', ['post' => $post, 'postId' => $id, 'tags' => $tags]);
@@ -52,6 +61,10 @@ class PostController extends Controller
 
     public function getAdminDelete($id)
     {
+        if(!Auth::check()){
+            return redirect()->back();
+        }
+
         $post = Post::find($id);
         $post->likes()->delete();
         $post->tags()->detach();
@@ -61,6 +74,10 @@ class PostController extends Controller
 
     public function postAdminCreate(Request $request)
     {
+        if(!Auth::check()){
+            return redirect()->back();
+        }
+
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10'
@@ -81,6 +98,9 @@ class PostController extends Controller
 
     public function postAdminUpdate(Request $request)
     {
+        if(!Auth::check()){
+            return redirect()->back();
+        }
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10'
